1. Заставили код компилироваться  - вынесли Solution в отдельный файл или был цикличный импорт. 2. Бобавили пример unit тестов
