+++ solver.py
@@ -2,8 +2,8 @@ from typing import List
 import numpy as np
 from listener import Listener
 from trial import Point
-from trial import Trial
 from problem import Problem
+from solution import Solution
 
 class SolverParameters:
     def __init__(self,
@@ -34,23 +34,7 @@ class SolverParameters:
         self.refineSolution = refineSolution
         self.startPoint = startPoint
 
-class Solution:
-    def __init__(self,
-                 problem: Problem,
-                 bestTrials: np.ndarray(shape = (1), dtype = Trial) = [],
 
-                 numberOfGlobalTrials: int = 0,
-                 numberOfLocalTrials: int = 0,
-                 solvingTime: np.double = 0.0,
-                 solutionAccuracy: np.double = 0.0
-                ):
-        self.problem = problem
-        self.bestTrials = bestTrials
-
-        self.numberOfGlobalTrials = numberOfGlobalTrials
-        self.numberOfLocalTrials = numberOfLocalTrials
-        self.solvingTime = solvingTime
-        self.solutionAccuracy = solutionAccuracy
 
 class Solver:
     __listeners: List[Listener] = []
